weekend works

+++ WindowsFormsApplication2/GeminiCore/Memory.cs
@@ -13,7 +13,7 @@ namespace GeminiCore
     public class Memory
     {
 
-        Dictionary<string, double> memoryBook = new Dictionary<string, double>();
+        Dictionary<string, Int16> memoryBook = new Dictionary<string, Int16>();
